Rework enemy handling
Choices were 1. an array for each type of enemy 2. a union of all enemy types or this option 3. a single class with all of the required fields.
Union makes no sense, as it will be at least as big as the largest enemy type anyway.
1. Just feels bad.
So 3 - there will be some wasted space, but can be minimised by choosing miminal number of data fields, reusing fields, etc.
Also, will it really make any difference? Unlikely.

Regardless, want to use a function pointer to determine behaviour as it should be nicely modular and flexible, whilst keeping things self contained.